requirements.txt
beautifulsoup4==4.11.1
Django==4.1.1
djangorestframework==3.14.0
requests==2.28.1
urllib3==1.26.12
